Update the format of the solution file to allow the source of extras … (#80)
* Update the format of the solution file to allow the source of extras to be preserved.

A new format in the explanation clause provides a [] delimited extra that supplied
the requirement. For example, if cryptography was provided by the "security" extra
from the requests library and your project required requests[security], the lines
would now appear as:

    myproject==1.0
    requests==2.0  # myproject (>=2 [security])
    cryptography==1.0.0  # requests[security] (>=1)

The extra marker is required to build the dependency tree back exactly. Additionally,
a new flag --remove-constraints will allow easy refreshing of solution files, by removing
the pins they provide when passed via --constraints, producing a solution with the same
versions but refreshed metadata.

This fixes #79

* Fixed formatting

---------

Co-authored-by: Andre Brisco <andre.brisco@protonmail.com>